Feat: enhance selecting
