+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.2",
+  "version": "0.8.2+1",
   "private": false,
   "scripts": {
     "dev": "next dev",
