+++ src/components/voucher/voucher_list.tsx
@@ -1,73 +1,12 @@
 import React, { useState } from 'react';
-import { BsFillTriangleFill } from 'react-icons/bs';
 import { MdOutlineFileDownload } from 'react-icons/md';
 import { FaRegTrashAlt } from 'react-icons/fa';
 import { Button } from '@/components/button/button';
 import VoucherItem from '@/components/voucher/voucher_item';
 import Pagination from '@/components/pagination/pagination';
+import SortingButton from '@/components/voucher/sorting_button';
 import { checkboxStyle } from '@/constants/display';
-
-enum SortOrder {
-  ASC = 'asc',
-  DESC = 'desc',
-}
-
-// Info: (20240920 - Julian) 排序按鈕
-const SortingButton = ({
-  string,
-  sortOrder,
-  setSortOrder,
-}: {
-  string: string;
-  sortOrder: null | SortOrder;
-  setSortOrder: (sortOrder: null | SortOrder) => void;
-}) => {
-  // Info: (20240920 - Julian) 初始無排序 -> 點擊後變成 ASC -> 再點擊變成 DESC -> 再點擊變回無排序
-  const clickHandler = () => {
-    switch (sortOrder) {
-      case null:
-        setSortOrder(SortOrder.ASC);
-        break;
-      case SortOrder.ASC:
-        setSortOrder(SortOrder.DESC);
-        break;
-      case SortOrder.DESC:
-        setSortOrder(null);
-        break;
-      default:
-        setSortOrder(null);
-        break;
-    }
-  };
-
-  return (
-    <button
-      id={`sorting-button-${string}`}
-      type="button"
-      onClick={clickHandler}
-      className="flex w-full items-center justify-center gap-4px"
-    >
-      {/* Info: (20240920 - Julian) 如果有選擇排序，則文字變成橙色 */}
-      <p className={sortOrder === null ? '' : 'text-text-brand-primary-lv1'}>{string}</p>
-      <div className="flex flex-col gap-px">
-        {/* Info: (20240920 - Julian) 向上箭頭：如果為升冪排序，則變成橙色 */}
-        <BsFillTriangleFill
-          size={8}
-          className={
-            sortOrder === SortOrder.ASC
-              ? 'text-text-brand-primary-lv1'
-              : 'text-surface-neutral-mute'
-          }
-        />
-        {/* Info: (20240920 - Julian) 向下箭頭：如果為降冪排序，則變成橙色 */}
-        <BsFillTriangleFill
-          size={8}
-          className={`rotate-180 ${sortOrder === SortOrder.DESC ? 'text-text-brand-primary-lv1' : 'text-surface-neutral-mute'}`}
-        />
-      </div>
-    </button>
-  );
-};
+import { SortOrder } from '@/constants/sort';
 
 const VoucherList = () => {
   const [isSelecting, setIsSelecting] = useState(false);
@@ -83,6 +22,7 @@ const VoucherList = () => {
   // Info: (20240920 - Julian) css string
   const tableCellStyles = 'table-cell text-center align-middle';
   const sideBorderStyles = 'border-r border-stroke-neutral-quaternary';
+  const checkStyle = `${isSelecting ? 'table-cell' : 'hidden'} text-center align-middle border-r border-stroke-neutral-quaternary`;
 
   const selectToggleHandler = () => setIsSelecting((prev) => !prev);
 
@@ -147,7 +87,7 @@ const VoucherList = () => {
   );
 
   const displayedVoucherList = Array.from({ length: 10 }, (_, i) => i + 1).map((i) => {
-    return <VoucherItem key={i} />;
+    return <VoucherItem key={i} isSelecting={isSelecting} />;
   });
 
   return (
@@ -160,7 +100,7 @@ const VoucherList = () => {
         {/* Info: (20240920 - Julian) ---------------- Table Header ---------------- */}
         <div className="table-header-group h-60px border-stroke-neutral-quaternary bg-surface-neutral-surface-lv1 text-sm text-text-neutral-tertiary">
           <div className="table-row">
-            <div className={`${tableCellStyles} ${sideBorderStyles}`}>
+            <div className={checkStyle}>
               <input type="checkbox" className={checkboxStyle} />
             </div>
             <div className={`${tableCellStyles} ${sideBorderStyles}`}>{displayedIssuedDate}</div>
